requirements.txt
python-telegram-bot
'python-telegram-bot[job-queue]'
pip
